refactor: Group Service 인터페이스 분리
